refactor: move player info benchmark to the desktop module in order to access extended info blocks
