+++ protocol/osrs-221-desktop/src/benchmarks/resources/net/rsprot/protocol/game/outgoing/info/huffman.dat
