+++ protocol/osrs-221-model/build.gradle.kts
@@ -1,7 +1,5 @@
 plugins {
     alias(libs.plugins.kotlin.jvm)
-    alias(libs.plugins.jmh)
-    alias(libs.plugins.allopen)
 }
 
 dependencies {
@@ -12,35 +10,4 @@ dependencies {
     implementation(projects.protocol)
     implementation(projects.protocol.osrs221Internal)
     implementation(projects.protocol.osrs221Shared)
-}
-
-allOpen {
-    annotation("org.openjdk.jmh.annotations.State")
-}
-
-sourceSets.create("benchmarks")
-
-kotlin.sourceSets.getByName("benchmarks") {
-    dependencies {
-        implementation(libs.jmh.runtime)
-        val mainSourceSet by sourceSets.main
-        val testSourceSet by sourceSets.test
-        val sourceSets = listOf(mainSourceSet, testSourceSet)
-        for (sourceSet in sourceSets) {
-            implementation(sourceSet.output)
-            implementation(sourceSet.runtimeClasspath)
-        }
-    }
-}
-
-benchmark {
-    targets {
-        register("benchmarks")
-    }
-
-    configurations {
-        register("PlayerInfoBenchmark") {
-            include("net.rsprot.protocol.game.outgoing.info.PlayerInfoBenchmark")
-        }
-    }
 }
